+++ 3DSkit.py
@@ -10,7 +10,7 @@ from util.fileops import bread, bwrite
 from util.help import main_help
 from util import error
 
-__version__ = '1.16.30'
+__version__ = '1.16.31'
 
 
 def parse_opts(s):